Continuing with switching from MMDB2, CMAPLIB and CLIPPER to GEMMI.
This commit now removes all references to FFTW2 and CLIPPER dependencies, as they are no longer required. The documentation and the README.md files have
been updated to reflect these changes, although the examples are yet to be re-run and the expected outputs re-computed - this is likely to be required as
Gemmi is now responsible for theoretical map computation and small differences in these are likely to cause small diferences in examples output.