Continuing with switching from MMDB2, CMAPLIB and CLIPPER to GEMMI.
Now the coding seems to be done, as ProSHADE can read and write both PDB and MAP files using Gemmi llibrary. The changes is, however, not yet complete as
the code needs to be checked for any remaining snippets of MMDB2, CMAPLIB, FFTW2 and CLIPPER codes, the installation needs to have the MMDB2 and CMAPLIB
removed, the examples needs to be re-computed and the documentation needs to be changed. However, the heavy lifting part is now done!