Finished with switching from MMDB2, CMAPLIB and CLIPPER to GEMMI.
The examples have now been re-tested and their expected results are now in line with the new code. The documentation and the README.md have now been updated to reflect the changes, as well as the standard dependencies, which now include the zlib required by gemmi.

This should now complete the switch to Gemmi.