Started adding support for Windows 10 compilation using Cygwin.
The initial steps required for compilation of ProSHADE under Windows 10 using Cygwin are now described in the standardDependencies.txt file. Following these will allow compiling and linking the proshade.exe and proshade.dll parts, but the python modules are yet to be supported under Windows.

The changes required to allow for this are mainly in the CMake files, where the order of the fftw3 library is important and was wrong before.
Also, some minor changes were done to remove the warnings resulting from comparing signed and unsigned ints.