This is the first commit in a new effort to improve the symmetry detection accuracy.
Generally, the idea is that by converting the rotation function to a spherical co-ordinates space,
where the radius is the angle-axis representation angle and the lattitude and longitude are the
angle-axis representation axis, searching for point groups (symmetry groups) would be easier. This
should be better in two major way, firstly by allowing simultaneous optimisation over all appeopriate
spheres at the same time (by searching for the axis which has the highest sum over all spheres) and
secondly by allowing simple searches in cases where the symmetry is know.

This commit adds the ability to map the rotation function onto the set of spheres as descrived above.
It also removes a bug in conversion of the SOFT map co-ordinates onto the Euler angles, where the Beta
angle was reported slightly off the correct value.