This is the second commit in a new effort to improve the symmetry detection accuracy.
This commit introduces multiple important changes. Firstly, it is now possible to selecte which of the two approaches (the original peak searching in rotation function space, or the new and now default angle-axis space) should be used - see the -z option. Consequently, this commit adds the ability to search the angle-axis space for a KNOW CYCLIC symmetry - the general symmetry search as well as the D, T, O and I symmetry searches are not yet available. Finally, this commit also adds the ability to use bi-cubic interpolation to allow for optimisation in the angle-axis space of the initial grid point search results.

This is still highly experimental and should for now stay in the experimental branch.