This is the third commit in a new effort to improve the symmetry detection accuracy.
This commit solves the issue of converting angle-axis values with axis 0;0;1 (i.e. Z axis), where the standard conversion leads to Euler angles of 0;0;0 regardless of the angle-axis angle.

The problem with this is that the Euler angles ZXZ conversion from rotation matrix does not use the rotation matrix elements 0;0 , 0;1 , 1;0 and 1;1 - all the elements required to code the Z axis rotation (as this is coded in the angle). The solution is in full search of all the rotation function indices and comparing their angle-axis value to the sought angle-axis value. This is quite time consuming, but there is a heuristic speed-up and it is used only in a fraction of cases, so overall it is not a big issue.