This is the tenth commit in a new effort to improve the symmetry detection accuracy.
This commit changes the peak IQR from median threshold from fixed value to linear dependence on resolution for angle-axis symmetry detection and the original value of 5.0 for distances and old symmetry detection algorithm, unless user specifies a value, in which case the threshold is that value.
It also applies the axis similarity threshold based on bandwidth, if the changing behaviour is allowed.

Both of these changes are required to make the detection work more accurately for larger range of resolutions.