This is the eleventh commit in a new effort to improve the symmetry detection accuracy.
This commit changes the cyclic point group detection algorithm from using the difference between angles to using prime numbers and detected groups combinations.

This commit also changes the thresholding to use non-peaks instead of peaks and several other inner workings to improve accuracy and timing.

Finally, this should be a somewhat final version that can now replace the old algorithm in the master branch in terms of C and D symmetries. T, O and I symmetries still need to be improved properly.