Another commit with progress towards the PyBind11 switch.
Now the whole ProSHADE_settings class is exported as well as a few ProSHADE_run class functions. The large success is with the getEnergyLevelsVector function, which is now exported to PyBind11 as a lambda function which automatically converts the identically named C++ functions return value of std::vector<double> to Numpy's float32 array and seamlessly gives this type to python.

More work now needs to be done on the ProSHADE_run class, but then the simpleAccess structures will all be converted to be using PyBind11.