Another commit regarding the PyBind11 switch.
Now, all of the simpleAccess functionality is exported to the python module. The advancedAccess and directAccess functions are not yet available, but they will be there soon enough.

Also, now I have added some examples of the new module. These are by no means complete or finished, but they serve as simple test cases for developing the functionality.

The SWIG codes are still left in, as I will do the final switch when I have a complete and fully working PyBind python module ready fot PyPi.