Another commit adding the PyBind11 bindings.
Now the new python module works with both, simpleAccess and advanceAccess examples. The directAccess still needs to be completed.

There are a few problems that need some attention. Firstly, it seems the overlay is slightly off and I need to have a look what is happening. Also, the PyBind11 examples are not will documented and writeen, so they will also need some attention.