Completed the directAccess.py example.
This means that multiple complex bindings needed to be added to allow direct access to rotation map, translation map, E matrices, SO(3) coefficients and other inner values. Also, this means that the PyBind11 bindings should now be complete.

Still, the SWIG codes need to be removed and the documentation needs to be completed for the advanced and direct access.